requirements.txt
numpy
pandas
tensorflow-cpu
gradio
pillow
nltk
tflearn
